media-source/URL-createObjectURL-revoke.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>insertAdjacentHTML</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div>
    <video id='v' controls autoplay width='320' height='240'></video>
</div>
<div id="log"></div>

<script>
var video = document.querySelector('video');
test(function() {
    var mediaSource = new MediaSource();
    var url = window.URL.createObjectURL(mediaSource);
    window.URL.revokeObjectURL(url);
    video.src = url;
    assert_equals(mediaSource.readyState, 'closed');
}, "URL.createObjectURL : 4. Revoke the MediaSource object URL by calling revokeObjectURL() on it.");
</script>
</body>
</html>
